third_party/blink/web_tests/platform/generic/fast/css/invalidation/target-pseudo-expected.txt

Use descendant invalidation set for :target pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(match, '').backgroundColor is transparent
PASS getComputedStyle(match, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
